fixed the data that comes back from the database to reflect the actual database requested
