+++ .env
@@ -4,7 +4,7 @@ USER_DB_USER_POSTGRES = 'xvcmlhle'
 USER_DB_PW_POSTGRES = '8K99q7Jrno3p41a95fll4u6EcM7B-fCC'
 # TOKEN_KEY = <any string>
 
-USER_DB_URL_MYSQL = database-1.cqvoaezvpbkr.us-east-2.rds.amazonaws.com
+USER_DB_URL_MYSQL = dbspytest.cqvoaezvpbkr.us-east-2.rds.amazonaws.com
 USER_DB_USER_MYSQL = 'admin'
 USER_DB_PW_MYSQL = 'Codesmith'
 TOKEN_KEY = <any string>
@@ -23,3 +23,9 @@ PG_TEST_PW = '8K99q7Jrno3p41a95fll4u6EcM7B-fCC'
 ## test user with saved schema to test save/load functionality
 TEST_USER_EMAIL = <email string>
 TEST_USER_PW = <password string>
+
+#                            this doesnt really matter            |----- only one from below ----|
+## URL must follow this format: ## databasetype://username:password@host name or server name or end point/database name ##
+
+# postgress url postgres://xvcmlhle:8K99q7Jrno3p41a95fll4u6EcM7B-fCC@mahmud.db.elephantsql.com/xvcmlhle
+# mysql url mysql://admin:codesmith@dbshy.cqvoaezvpbkr.us-east-2.rds.amazonaws.com/user
