+++ src/components/DBDisplay/Sidebar.tsx
@@ -51,6 +51,8 @@ const Sidebar = (props: any) => {
       }
     }
 
+console.log('values: ', values)
+
     //update dbCredentials
     setDbCredentials(values);
     setConnectPressed(true);
@@ -95,6 +97,7 @@ const Sidebar = (props: any) => {
         >
           <option value="postgres">PostgreSQL</option>
           <option value="mysql">MySQL</option>
+          <option value="mssql">Microsoft SQL</option>
         </select>
       </span>
       <br></br>
